% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{Coll_Blasco_2006}
\alias{Coll_Blasco_2006}
\title{Data: Coll and Blasco (2006).}
\format{
Data frame with 6 rows and 5 columns. Definition of inputs (X) and outputs (Y):
\describe{
  \item{x1 = Employees}{Number of employees}
  \item{x2 = Capital}{Impairment of assets}
  \item{y1 = Vehicles}{Number of vehicles sold}
  \item{y2 = Orders}{Number of orders received at the garage}
}
}
\source{
Coll-Serrano, V.; Blasco-Blasco, O. (2006). Evaluacion de la Eficiencia
mediante el Análisis Envolvente de Datos. Introduccion a los Modelos Básicos.
}
\usage{
data("Coll_Blasco_2006")
}
\description{
Data of six authorized dealers with two inputs and two outputs.
}
\examples{
# Example. How to read data with deaR
data("Coll_Blasco_2006")
data_example <- make_deadata(Coll_Blasco_2006,
                             dmus = 1,
                             ni = 2,
                             no = 2)

}
\seealso{
\code{\link{make_deadata}}
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolos} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benitez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
\keyword{datasets}
