% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{efficiencies}
\alias{efficiencies}
\title{Efficiencies}
\usage{
efficiencies(x, ...)
}
\arguments{
\item{x}{dea / deafuzzy object}

\item{...}{ignored}
}
\description{
Extracts efficiencies from  dea/dea_fuzzy objects.
}
