% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dea_fuzzy.R
\name{is.dea_fuzzy}
\alias{is.dea_fuzzy}
\title{dea_fuzzy class check.}
\usage{
is.dea_fuzzy(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a dea_fuzzy object (that is, has "dea_fuzzy" 
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks wether an R object is of dea_fuzzy class or not.
}
