% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.deadata_fuzzy.R
\name{is.deadata_fuzzy}
\alias{is.deadata_fuzzy}
\title{deadata_fuzzy class check.}
\usage{
is.deadata_fuzzy(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a deadata_fuzzy object (that is, has "deadata_fuzzy" 
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks wether an R object is of deadata_fuzzy class or not.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
