% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_profit.R
\name{model_profit}
\alias{model_profit}
\title{Profit efficiency DEA model.}
\usage{
model_profit(datadea,
            dmu_eval = NULL,
            dmu_ref = NULL,
            price_input = NULL,
            price_output = NULL,
            rts = c("crs", "vrs", "nirs", "ndrs", "grs"),
            L = 1,
            U = 1,
            restricted_optimal = TRUE,
            returnlp = FALSE,
            ...)
}
\arguments{
\item{datadea}{The data, including \code{n} DMUs, \code{m} inputs and \code{s} outputs.}

\item{dmu_eval}{A numeric vector containing which DMUs have to be evaluated.}

\item{dmu_ref}{A numeric vector containing which DMUs are the evaluation reference set.}

\item{price_input}{Unit prices of inputs for cost or profit efficiency models. It is a value, vector of length \code{m}, or matrix \code{m} x \code{ne} (where \code{ne} is the lenght of \code{dmu_eval}).}

\item{price_output}{Unit prices of outputs for revenue or profit efficiency models. It is a value, vector of length \code{s}, or matrix \code{s} x \code{ne}.}

\item{rts}{A string, determining the type of returns to scale, equal to "crs" (constant),
"vrs" (variable), "nirs" (non-increasing), "ndrs" (non-decreasing) or "grs" (generalized).}

\item{L}{Lower bound for the generalized returns to scale (grs).}

\item{U}{Upper bound for the generalized returns to scale (grs).}

\item{restricted_optimal}{Logical. If it is \code{TRUE}, the optimal inputs are restricted to be <= inputs (for cost efficiency models) or
the optimal outputs are restricted to be >= outputs (for revenue efficiency models).}

\item{returnlp}{Logical. If it is \code{TRUE}, it returns the linear problems (objective function and constraints) of stage 1.}

\item{...}{Ignored, for compatibility issues.}
}
\description{
Cost, revenue and profit efficiency DEA models.
}
\examples{
 data("Fortune500")
 data_cost <- read_data(datadea = Fortune500,
                        ni = 3, 
                        no = 2)
 result <- model_profit(data_cost, 
                        price_input = 1)
 efficiencies(result)
 
}
\references{
Zhu, J. (2014). Quantitative Models for Performance Evaluation and Benchmarking. Data Envelopment Analysis with Spreadsheets. 3rd Edition Springer, New York. DOI: 10.1007/978-3-319-06647-9
}
\seealso{
\code{\link{model_deaps}}, \code{\link{model_nonradial}}, \code{\link{model_sbmeff}}
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
