% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dea.R
\name{plot.dea}
\alias{plot.dea}
\title{Plot}
\usage{
\method{plot}{dea}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dea"} obtained by a dea model function.}

\item{...}{Ignored, for compatibility issues.}
}
\value{
Depending on the model it returns a single data.frame containing: efficiencies, 
slacks, lambdas, targets, references or a list of data.frames with the cross-efficiencies computed 
with different methods (Arbitrary, Method II or Method III (see CITA)) or, in case the model is a
 malmquist index, a single data.frame with the coefficients for the different periods.
}
\description{
Plot some attribute of a dea model
}
\examples{
data_example <- read_data(datadea = Fortune500, dmus = 1, inputs = 2:4, outputs = 5:6)

deamodel <- model_basic(data_example)
plot(deamodel)
}
