% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.deafuzzy.R
\name{plot.dea_fuzzy}
\alias{plot.dea_fuzzy}
\title{Plot for Fuzzy DEA models.}
\usage{
\method{plot}{dea_fuzzy}(x, showPlots = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"dea_fuzzy"} obtained by a fuzzy dea model function.}

\item{showPlots}{Logical. When TRUE (default) the plots are shown one by one. When it 
is FALSE the plots are not shown and are returned by the function (invisiblily) as a 
list.}

\item{...}{Ignored, for compatibility issues.}
}
\value{
Depending on the model it returns ...
}
\description{
Plot some attribute of a Fuzzy DEA model (Guo-Tanaka and Kao-Liu models).
}
\examples{


}
\references{
#' Zhu, J. (2014). Quantitative Models for Performance Evaluation and Benchmarking. Data Envelopment Analysis with Spreadsheets. 3rd Edition Springer, New York. DOI: 10.1007/978-3-319-06647-9
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
