% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.deadata.R
\name{print.deadata}
\alias{print.deadata}
\title{deadata class print method}
\usage{
\method{print}{deadata}(x, ...)
}
\arguments{
\item{x}{A \code{deadata} object (as returned by \code{make_deadata} function).}

\item{...}{For compatibility issues.}
}
\description{
Print method for \code{deadata} class.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
