% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{Retails}
\alias{Retails}
\title{Data: Norman and Stoker (1991).}
\format{Data frame with 45 rows and 20 columns. Definition of variables:

  V1 = Staff costs
  
  V2 = Non-staff controlable cost
  
  V3 = Controllable cost
  
  V4 = Uncontrollable cost
  
  V5 = Non-staff cost
  
  V6 = Total cost
  
  V7 = Revenue Gp1
  
  V8 = Revenue Gp2
  
  V9 = Revenue Gp3
  
  V10 = Total revenue
  
  V11 = Revenue growth
  
  V12 = Revenue % growth
  
  V13 = Central overheads
  
  V14 = Profit
  
  V15 = Pitch
  
  V16 = Age of shop (in months)
  
  V17 = Catchment population
  
  V18 = Competititon
  
  V19 = Customer service}
\source{
Norman, M. and Stoker, B. (1991). Data Envelopment Analysis. The Assessment of Performance. New York: John Wiley & Sons, Inc.
}
\usage{
data("Retails")
}
\description{
Data of 45 shops that refer to 19 variables.
}
\examples{
# Example 1. Replications of results in Norman and Stocker (1991, p.59)
data("Retails")
data_example <- read_data(Retails,
                          dmus=1, 
                          inputs=7, 
                          outputs=c(11,12))
result <- model_multiplier(data_example, 
                           orientation="io", 
                           rts="crs")
score <- efficiencies(result)
weights <- multipliers(result)
eff <- data.frame(score,weights$input,weights$output)
eff

# Example 2. Replications of results in Norman and Stocker (1991, p.114-115)
data("Retails")
data_example2 <- read_data(Retails,
                           dmus=1, 
                           inputs=11, 
                           outputs=12)
result2 <- model_basic(data_example2,
                       orientation="io", 
                       rts="vrs", 
                       compute_multiplier=TRUE)
score2 <- efficiencies(result2)
references <- references(result2)
score2
references

}
\seealso{
\code{\link{read_data}}, \code{\link{model_basic}}
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
\keyword{datasets}
