% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barlab_order.R
\name{barlab_order}
\alias{barlab_order}
\title{barplot with label and sorted frequency}
\usage{
barlab_order(x, percent = FALSE, title = deparse(substitute(x)),
  unlimit = FALSE)
}
\arguments{
\item{x}{a categorical variable.}

\item{percent}{a logical for display label as percentage.}

\item{title}{a title of plot}

\item{unlimit}{a logical for set plot more than 10 categories}
}
\description{
barplot with label and sorted frequency
}
\examples{
a <- c("A", "A", "B", "B", "B", "C")
barlab_order(a)
barlab_order(a, percent = TRUE)
}
