% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_diag.R
\name{dist_diag}
\alias{dist_diag}
\title{create histogram and boxplot of a numeric variable}
\usage{
dist_diag(x, title = deparse(substitute(x)))
}
\arguments{
\item{x}{a numerical vector}

\item{title}{title of plot}
}
\value{
histogram and box plot
}
\description{
create histogram and boxplot of a numeric variable
}
\examples{
dist_diag(mtcars$mpg)
}
