% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{efficiencies}
\alias{efficiencies}
\title{Objective scores}
\usage{
efficiencies(x, ...)
}
\arguments{
\item{x}{DEA / DEA fuzzy object}

\item{...}{ignored}
}
\description{
Extract the scores (optimal objective values) of the evaluated
DMUs from a DEA / DEA fuzzy solution. Note that these scores may not always
be interpreted as efficiencies.
}
