% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{efficiencies}
\alias{efficiencies}
\title{Objective scores}
\usage{
efficiencies(x, ...)
}
\arguments{
\item{x}{An object of class \code{dea}, \code{dea_fuzzy} or \code{dea_stoch}.}

\item{...}{ignored.}
}
\description{
Extract the scores (optimal objective values) of the evaluated
DMUs from a conventional, fuzzy or stochastic DEA solution. Note that these
scores may not always be interpreted as efficiencies.
}
