% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielab.R
\name{pielab}
\alias{pielab}
\title{pielab}
\usage{
pielab(x, lb = c("Cat1", "Cat2"), cl = c("cyan", "pink"), tt = "Title",
  center = FALSE, fade = FALSE)
}
\arguments{
\item{x}{a vector of categories variable}

\item{lb}{a label of each category}

\item{cl}{a vector of colors}

\item{tt}{a title of plot}

\item{center}{start division line at center}

\item{fade}{make the pie chart fade}
}
\value{
piechar and freqeuncy_table
}
\description{
create a pie chart with percentage label.
}
\examples{
pielab(iris$Species)
}
