% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielab.R
\name{pielab}
\alias{pielab}
\title{pielab}
\usage{
pielab(x, label = "default", col = "default",
  title = deparse(substitute(x)), center = FALSE, fade = FALSE)
}
\arguments{
\item{x}{a vector of categories variable}

\item{label}{a label of each category}

\item{col}{a vector of colors}

\item{center}{start division line at center}

\item{fade}{make the pie chart fade}

\item{titile}{a title of plot}
}
\value{
piechar and freqeuncy_table
}
\description{
create a pie chart with percentage label.
}
\examples{
a <- c("A","A","A","B","B","C")
pielab(a)
}
