% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(datadea = NULL,
          ni = NULL,
          no = NULL,
          dmus = 1,
          inputs = NULL,
          outputs = NULL,
          nc_inputs = NULL,
          nc_outputs = NULL,
          nd_inputs = NULL,
          nd_outputs = NULL,
          ud_inputs = NULL, 
          ud_outputs = NULL)
}
\arguments{
\item{datadea}{Data frame with DEA data.}

\item{ni}{Number of inputs, if inputs are in columns 2:(\code{ni} + 1) (if DMUs
are in the first column) or 1:\code{ni} (no DMUs column).}

\item{no}{Number of outputs, if outputs are in columns (\code{ni} + 2):(\code{ni} +
\code{no} + 1) (if DMUs are in the first column) or (\code{ni} + 1):(\code{ni} +
\code{no}) (no DMUs column). If not specified, DMUs are in the first column.}

\item{dmus}{Column (number or name) of DMUs (optional). By default, it is the
first column. If there is not any DMU column, then it must be \code{NULL}.}

\item{inputs}{Columns (numbers or names) of inputs (optional). It prevails over \code{ni}.
Alternatively to \code{datadea}, it can be a matrix with the inputs (DMUs in columns).
In this case, DMUs names are taken from the columns names.}

\item{outputs}{Columns (numbers or names) of outputs (optional). It prevails over \code{no}.
Alternatively to \code{datadea}, it can be a matrix with the outputs (DMUs in columns).}

\item{nc_inputs}{A numeric vector containing the indices of non-controllable inputs.}

\item{nc_outputs}{A numeric vector containing the indices of non-controllable outputs.}

\item{nd_inputs}{A numeric vector containing the indices of non-discretionary inputs.}

\item{nd_outputs}{A numeric vector containing the indices of non-discretionary outputs.}

\item{ud_inputs}{A numeric vector containing the indices of undesirable (good) inputs.}

\item{ud_outputs}{A numeric vector containing the indices of undesirable (bad) outputs.}
}
\description{
This function is deprecated. Use \code{make_deadata} instead.
}
