% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_malmquist.R
\name{read_malmquist}
\alias{read_malmquist}
\title{read_malmquist}
\usage{
read_data(datadea, 
             nper, 
             percol, 
             arrangement = c("horizontal", "vertical"),
             ...)
}
\arguments{
\item{datadea}{Data frame with DEA data.}

\item{nper}{Number of time periods (with dataset in wide format).}

\item{percol}{Column of time period (with dataset in long format)}

\item{ni}{Number of inputs, if inputs are in columns 2:(\code{ni} + 1) (if DMUs are in the first column) or 1:\code{ni} (no DMUs column).}

\item{no}{Number of outputs, if outputs are in columns (\code{ni} + 2):(\code{ni} + \code{no} + 1) (if DMUs are in the first column) or (\code{ni} + 1):(\code{ni} + \code{no}) (no DMUs column).}

\item{dmus}{Column (number or name) of DMUs (optional). If there is not any DMU column, then it must be NULL. If not specified, DMUs are in the first column.}

\item{inputs}{Columns (numbers or names) of inputs (optional). It prevails over \code{ni}.}

\item{outputs}{Columns (numbers or names) of outputs (optional). It prevails over \code{no}.}

\item{nc_inputs}{A numeric vector containing non-controllable inputs.}

\item{nc_outputs}{A numeric vector containing non-controllable outputs.}

\item{nd_inputs}{A numeric vector containing non-discretionary inputs.}

\item{nd_outputs}{A numeric vector containing non-discretionary outputs.}
}
\value{
An object of class \code{deadata}
}
\description{
This function creates, from a data frame, a list \code{deadata},
}
\examples{
# Example 1. If you have a dataset in wide format.
data("Economy")
data_example <- read_malmquist(datadea = Economy, 
                               nper = 5, 
                               arrangement = "horizontal",
                               ni = 2, 
                               no = 1)
# This is the same as:
data_example <- read_malmquist(datadea = Economy,
                               nper = 5, 
                               arrangement = "horizontal",
                               inputs = 2:3, 
                               outputs = 4)
# Example 2. If you have a dataset in long format.
data("EconomyLong")
data_example2 <- read_malmquist(EconomyLong,
                                percol = 2, 
                                arrangement = "vertical",
                                inputs = 3:4, 
                                outputs = 5)

}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
