% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dea.R
\name{summary.dea}
\alias{summary.dea}
\title{Summary conventional DEA model.}
\usage{
\method{summary}{dea}(object, exportExcel = TRUE, filename = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"dea"} obtained by a dea model function.}

\item{exportExcel}{Logical value. If TRUE (default) the results are also exported to an Excel file}

\item{filename}{Character string. Absolute filename (including path) of the exported Excel file. 
If NULL, then the name of the file will be "ResultsDEA"+timestamp.xlsx.}

\item{...}{Ignored. Used for compatibility issues.}
}
\value{
Depending on the model it returns a single data.frame containing: efficiencies, 
slacks, lambdas, targets, references or a list of data.frames with the cross-efficiencies computed 
with different methods (Arbitrary, Method II or Method III (see CITA)) or, in case the model is a
 malmquist index, a single data.frame with the coefficients for the different periods.
}
\description{
Summary of the results obtained by a conventiona DEA model.
}
\examples{
data("PFT1981") 
# Selecting DMUs in Program Follow Through (PFT)
PFT <- PFT1981[1:49, ] 
PFT <- read_data(PFT, 
                 dmus = 1, 
                 inputs = 2:6, 
                 outputs = 7:9 )
eval_pft <- model_basic(PFT, 
                        orientation = "io", 
                        rts = "crs")
summary(eval_pft, exporExcel = FALSE)
}
